Calculate new value
Close #31
