+++ codevance_api/payments/tests/conftest.py
@@ -1,8 +1,10 @@
+from datetime import date, timedelta
+
 import pytest
 from django.contrib.auth import get_user_model
 from rest_framework.test import APIClient
 
-from codevance_api.payments.models import Supplier
+from codevance_api.payments.models import Supplier, Payment
 
 
 @pytest.fixture
@@ -51,3 +53,29 @@ def auth_supplier_01(db, supplier_01):
     client = APIClient()
     client.force_authenticate(user=supplier_01.user)
     return client
+
+
+@pytest.fixture
+def payment_supplier_01(supplier_01):
+    """
+    Creates and returns a payment that
+    belongs to supplier 01.
+    """
+    due_date = date.today() + timedelta(days=10)
+    payment = Payment.objects.create(supplier=supplier_01,
+                                     due_date=due_date,
+                                     value=1100)
+    return payment
+
+
+@pytest.fixture
+def payment_supplier_02(supplier_02):
+    """
+    Creates and returns a payment that
+    belongs to supplier 02.
+    """
+    due_date = date.today() + timedelta(days=10)
+    payment = Payment.objects.create(supplier=supplier_02,
+                                     due_date=due_date,
+                                     value=1200)
+    return payment
