+++ codevance_api/payments/tests/test_new_value_calculation.py
@@ -1,0 +1,12 @@
+from datetime import timedelta
+
+from codevance_api.payments.payments import calc_new_value
+
+
+def test_calc_new_value(payment_supplier_01):
+    """
+    Certifies that the new value is calculated correctly.
+    """
+    new_due_date = payment_supplier_01.due_date + timedelta(days=15)
+    new_value = calc_new_value(payment_supplier_01, new_due_date)
+    assert new_value == 1083.5
